New 0.6.1-2 version

+++ ps-build-cs.sh
@@ -4,7 +4,7 @@ cd $(dirname $0)
 source ps-params.sh
 
 # clean
-cd $PS_BUILD/cs
+cd "$PS_BUILD/cs"
 jam clean
 
 # update
@@ -14,20 +14,42 @@ then
 fi
 
 # point to Cal3D and Bullet
-export LD_LIBRARY_PATH=$PS_BUILD/cal3d/src/cal3d/.libs/:$PS_BUILD/bullet/lib/:$LD_LIBRARY_PATH
+export LD_LIBRARY_PATH="$PS_BUILD/cal3d/src/cal3d/.libs/:$PS_BUILD/bullet/lib/:$LD_LIBRARY_PATH"
 
-# use old gcc for CrystalSpace 3D
-echo '1' > ~choice.txt 
-sudo update-alternatives --config gcc < ~choice.txt
-rm -f ~choice.txt
+# get current gcc version (assume it's the latest gcc version available)
+GCC_VERSION=$(gcc --version | grep ^gcc | sed 's/^.* //g')
+GCC_VERSION=${GCC_VERSION:0:3}
+
+if [ "$GCC_VERSION" = "4.7" ]
+then
+	# install gcc 4.6, needed to build CrystalSpace 3D
+	sudo apt-get --assume-yes install gcc-4.6 g++-4.6
+
+	# setup gcc alternatives
+	sudo update-alternatives --remove-all gcc
+	sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-4.6 40 --slave /usr/bin/g++ g++ /usr/bin/g++-4.6
+	sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-4.8 60 --slave /usr/bin/g++ g++ /usr/bin/g++-4.8
+
+	# use gcc 4.6 for CrystalSpace 3D
+	echo '1' > ~choice.txt 
+	sudo update-alternatives --config gcc < ~choice.txt
+	rm -f ~choice.txt
+fi
 
 # build
-cd $PS_BUILD/cs
-./configure --without-java --without-perl --without-python --without-3ds --with-cal3d=$PS_BUILD/cal3d --with-bullet=$PS_BUILD/bullet
-jam -j3 -aq libs plugins cs-config # walktest
+cd "$PS_BUILD/cs"
+./configure --without-java --without-perl --without-python --without-3ds --with-cal3d="$PS_BUILD/cal3d" --with-bullet="$PS_BUILD/bullet"
+jam -j$CONCURRENT_JOBS -aq libs plugins cs-config # walktest
 
-# use "new" gcc for the rest
-echo '2' > ~choice.txt 
-sudo update-alternatives --config gcc < ~choice.txt
-rm -f ~choice.txt
+if [ "$GCC_VERSION" = "4.7" ]
+then
+	# use default (new) gcc for the rest
+	echo '0' > ~choice.txt 
+	sudo update-alternatives --config gcc < ~choice.txt
+	rm -f ~choice.txt
+	
+	# remove gcc alternatives
+	sudo update-alternatives --remove-all gcc
+fi
+
